Rename to globalThis
