+++ packages/jss/src/utils/globalThis.js
